// src/types/supabase-generated.ts
export type Database = {
  // Allows to automatically instanciate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "12.2.3 (519615d)"
  }
  public: {
    Tables: {
// ...
      profiles: {
        Row: {
          bio: string | null
          created_at: string | null
          elo_rating: number | null
          matches_played: number | null
          matches_won: number | null
          player_style_analysis: string | null
          profile_picture_url: string | null
          skill_level: string | null
          updated_at: string | null
          user_id: string
          username: string
        }
        Insert: {
          bio?: string | null
          created_at?: string | null
          elo_rating?: number | null
          matches_played?: number | null
          matches_won?: number | null
          player_style_analysis?: string | null
          profile_picture_url?: string | null
          skill_level?: string | null
          updated_at?: string | null
          user_id: string
          username: string
        }
        Update: {
          bio?: string | null
          created_at?: string | null
          elo_rating?: number | null
          matches_played?: number | null
          matches_won?: number | null
          player_style_analysis?: string | null
          profile_picture_url?: string | null
          skill_level?: string | null
          updated_at?: string | null
          user_id?: string
          username?: string
        }
        Relationships: []
      }
// ...
    }
// ...
  }
}